* develop: (59 commits)
  Refactored and cleaned up UI
  Added missing configyration for cean up job
  Added MIT license to project
  Model updated to work with MSSQL server
  Translates db script from MySQL to MSSQL
  Added comment to getactivehelpdesks unit tests
  Fixed/added most items detailed under bad practice issues in the code review
  Added unit tests to verify changes made to check-in method in facade and data layer
  Fixed most bug causing problems mentioned in code review + added thing suggested to avoid bug causing issues
  Added functionality to force checkout and remove queue items.
  improved password setting security
  Added edit user functionality
  Modified returned status code for UpdateUser
  varous changes
  Added missing documentation
  Added missing tables to database export job
  Finished UpdateQueueItem and UpdateQueueItemStatus. Added unit tests.
  Added check in, check out and delete timespan functionalities, fixed other little things (misspellings, using orders, etc...)
  Finished v1 of export database job
  Completed 'Add User' feature
  ...

+++ HelpdeskSystem/Helpdesk.Services/UsersFacade.cs
@@ -33,7 +33,7 @@ namespace Helpdesk.Services
         }
 
         /// <summary>
-        /// This method is responsible for retrieving all users from the database
+        /// This method is responsible for retrieving all users from the helpdesk system
         /// </summary>
         /// <returns>The response that indicates if the operation was a success,
         /// and the list of users</returns>
@@ -48,11 +48,6 @@ namespace Helpdesk.Services
                 var dataLayer = new UsersDataLayer();
 
                 List<UserDTO> users = dataLayer.GetUsers();
-
-                if (users.Count == 0)
-                {
-                    throw new NotFoundException("No users found!");
-                }
 
                 response.Users = users;
                 response.Status = HttpStatusCode.OK;
@@ -73,7 +68,7 @@ namespace Helpdesk.Services
         }
 
         /// <summary>
-        /// This method is responsible for getting a specific user from the database
+        /// This method is responsible for getting a specific user from the helpdesk system
         /// </summary>
         /// <param name="id">The UserId of the specific user to be retrieved</param>
         /// <returns>The response that indicates if the operation was a success,
@@ -130,13 +125,19 @@ namespace Helpdesk.Services
                 {
                     return response;
                 }
+
+                if (string.IsNullOrEmpty(request.Password))
+                    request.Password = request.Username;
+
                 request.Password = HashText(request.Password);
 
                 var dataLayer = new UsersDataLayer();
 
                 if (dataLayer.GetUserByUsername(request.Username) != null)
                 {
-                    throw new Exception("Unable to add user! User already exists!");
+                    response.Status = HttpStatusCode.Forbidden;
+                    response.StatusMessages.Add(new StatusMessage(HttpStatusCode.Forbidden, "Username already exists"));
+                    return response;
                 }
 
                 int? result = dataLayer.AddUser(request);
@@ -174,16 +175,15 @@ namespace Helpdesk.Services
             try
             {
                 response = (UpdateUserResponse)request.CheckValidation(response);
+
                 if (response.Status == HttpStatusCode.BadRequest)
-                {
                     return response;
-                }
 
                 request.Password = HashText(request.Password);
 
                 var dataLayer = new UsersDataLayer();
 
-                if (dataLayer.GetUserByUsername(request.Username) != null)
+                if (dataLayer.GetUserByUsername(request.Username)!=null && dataLayer.GetUserByUsername(request.Username).UserId != id)
                 {
                     throw new Exception("Unable to update user! User with username " + request.Username + "already exists!");
                 }
@@ -193,7 +193,7 @@ namespace Helpdesk.Services
                 if (result == false)
                     throw new NotFoundException("Unable to find user!");
 
-                response.result = result;
+                response.Result = result;
                 response.Status = HttpStatusCode.OK;
 
             }
@@ -206,8 +206,8 @@ namespace Helpdesk.Services
             catch (Exception ex)
             {
                 s_logger.Error(ex, "Unable to update user!");
-                response.Status = HttpStatusCode.InternalServerError;
-                response.StatusMessages.Add(new StatusMessage(HttpStatusCode.InternalServerError, "Unable to update user!"));
+                response.Status = HttpStatusCode.Forbidden;
+                response.StatusMessages.Add(new StatusMessage(HttpStatusCode.Forbidden, "Unable to update user!"));
             }
             return response;
         }
@@ -217,13 +217,22 @@ namespace Helpdesk.Services
         /// </summary>
         /// <param name="id">The id of the user to be deleted</param>
         /// <returns>A response that indicates whether or not the deletion was successful</returns>
-        public DeleteUserResponse DeleteUser(int id)
+        public DeleteUserResponse DeleteUser(int id, string currentUser)
         {
             var response = new DeleteUserResponse();
 
             try
             {
                 var dataLayer = new UsersDataLayer();
+
+                UserDTO user = dataLayer.GetUser(id);
+
+                if (user.Username == currentUser)
+                {
+                    response.Status = HttpStatusCode.Forbidden;
+                    return response;
+                }
+
                 bool result = dataLayer.DeleteUser(id);
 
                 if (result)
@@ -305,6 +314,14 @@ namespace Helpdesk.Services
                 var token = tokenHandler.WriteToken(rawToken);
 
                 response.Token = token;
+
+                if (user.FirstTime)
+                {
+                    response.Status = HttpStatusCode.Accepted;
+                    response.UserId = user.UserId;
+                    return response;
+                }
+
                 response.Status = HttpStatusCode.OK;
             }
             catch (Exception ex)
@@ -342,7 +359,7 @@ namespace Helpdesk.Services
 
                 UserDTO userFromUsername = dataLayer.GetUserByUsername(username);
 
-                if (!(userFromID.UserId == userFromUsername.UserId && userFromID.Username == userFromUsername.Username))
+                if (!(userFromID.UserId == userFromUsername.UserId && userFromID.Username == userFromUsername.Username && (!userFromID.FirstTime)))
                 {
                     s_logger.Warn("Unable to verify user.");
                     result = false;
